// src/types/gotohome.ts
/**
 * Member information types for travel insurance
 */

export type UsersInfo = {
  /** Full customer name */
  customer_name: string;
  /** Customer name in Nepali */
  customer_name_nep: string;
  /** First name */
  customer_first_name: string;
  /** Middle name */
  customer_middle_name: string;
  /** Last name */
  customer_last_name: string;
  /** First name in Nepali */
  customer_first_name_nep: string;
  /** Middle name in Nepali */
  customer_middle_name_nep: string;
  /** Last name in Nepali */
  customer_last_name_nep: string;
  /** Mobile phone number */
  mobile_no: string;
  /** KYC verification status */
  kyc_status: string;
}